Add styler class
