+++ sticknotes-frontend/src/app/board/board.component.html
@@ -1,12 +1,12 @@
 <div class="flex-container" *ngIf="board">
   <h3>{{board.title}}</h3>
-  <div class="rc-wrapper" [style]="getRCWrapperStyle()">
-    <div #columnsDiv class="horizontal-scroll" [style]="getBoardWrapperWidth()">
+  <div class="rc-wrapper" [style]="styler.getRCWrapperStyle()">
+    <div #columnsDiv class="horizontal-scroll" [style]="styler.getBoardWrapperWidth()">
       <!-- Scroll container for the visible width of board -->
-      <div class="c-wrapper" [style]="getBoardWidth()">
+      <div class="c-wrapper" [style]="styler.getBoardWidth()">
         <!-- Full width of board-->
         <div class="column-name" *ngFor="let columnName of boardColumnNames"
-          [style]="getColumnNameDivStyle(columnName)">
+          [style]="styler.getColumnNameDivStyle(columnName)">
           <div class="cn-flex-container">
             <p>{{columnName.title}}</p>
             <button class="note-button" mat-icon-button [matMenuTriggerFor]="menu">
@@ -14,20 +14,20 @@
             </button>
           </div>
         </div>
-        <div class="column-name" *ngFor="let coord of newColumnNameButtonCoordinates" [style]="getPlusButtonStyle(coord)">
+        <div class="column-name" *ngFor="let coord of newColumnNameButtonCoordinates" [style]="styler.getPlusButtonStyle(coord)">
           <button class="note-button" mat-icon-button (click)="openNewColumnDialog(coord)">
             <mat-icon class="note-icon">add</mat-icon>
           </button>
         </div>
       </div>
     </div>
-    <div class="r-wrapper" [style]="getBoardWrapperHeight()">
+    <div class="r-wrapper" [style]="styler.getBoardWrapperHeight()">
 
     </div>
-    <div #boardWrapper class="board-wrapper" [style]="getBoardWrapperStyle()" (scroll)="boardScrolled($event)">
-      <div class="board" [style]="getBoardStyle()">
+    <div #boardWrapper class="board-wrapper" [style]="styler.getBoardWrapperStyle()" (scroll)="boardScrolled($event)">
+      <div class="board" [style]="styler.getBoardStyle()">
         <div *ngFor="let y of board.rows | rangeLoop">
-          <div *ngFor="let x of board.cols | rangeLoop" class="new-note-container" [style]="getSlotStyle(x, y)">
+          <div *ngFor="let x of board.cols | rangeLoop" class="new-note-container" [style]="styler.getSlotStyle(x, y)">
             <button mat-icon-button aria-label="Create new note" (click)="openNewNoteDialog(x, y)">
               <mat-icon color="primary">add</mat-icon>
             </button>
@@ -37,7 +37,7 @@
       <div *ngFor="let note of board.notes; let i = index">
         <div cdkDragBoundary=".board" (cdkDragStarted)="onNoteDragStart($event)"
           (cdkDragEnded)="onNoteDrop($event, note)" cdkDrag [cdkDragDisabled]="!canModifyNote(note)"
-          class="mat-card-wrapper" [style]="getNoteStyle(note)">
+          class="mat-card-wrapper" [style]="styler.getNoteStyle(note)">
           <mat-card [style]="'background-color:' + note.color">
             <mat-card-content>
               <div class="card-header">
